Separate remaining Rust test suites
Move large inline test modules beside their production modules without changing runtime behavior.
